Depend

+++ scripts/1_process_sample.sh
@@ -3,7 +3,7 @@
 #SBATCH --job-name=1_process_sample
 #SBATCH --mem=64GB
 #SBATCH --cpus-per-task=4
-#SBATCH --time=12:00:00
+#SBATCH --time=2-00:00:00
 #SBATCH --partition=cgawad
 
 START_TIME=$(date +%s)
@@ -11,11 +11,12 @@ FASTQ_DIR=$1
 RESULTS_DIR=$2
 R1_SUFFIX=$3
 R2_SUFFIX=$4
-REF_FASTA=$5
-KRAKEN_DB_TYPES_ARRAY=( $(echo $6 | sed 's/-/ /g') )
-KRAKEN_DB_DIR_PREFIX=$7
-TOOLS_DIR=$8
-SAMPLE_ARRAY=( $(echo $9 | sed 's/:/ /g') )
+SKIP_TRIMMOMATIC=$5
+REF_FASTA=$6
+KRAKEN_DB_TYPES_ARRAY=( $(echo $7 | sed 's/-/ /g') )
+KRAKEN_DB_DIR_PREFIX=$8
+TOOLS_DIR=$9
+SAMPLE_ARRAY=( $(echo ${10} | sed 's/:/ /g') )
 SAMPLE=${SAMPLE_ARRAY[$(( $SLURM_ARRAY_TASK_ID - 1 ))]}
 
 echo -e "START: $(date)\nBacteria Pipeline\nFastq dir: $FASTQ_DIR\nResults dir: $RESULTS_DIR\nSample: $SAMPLE"
@@ -24,33 +25,39 @@ cd $RESULTS_DIR
 ml python/3.6.1 java 
 ml biology bwa samtools gatk
 
+R1_FASTQ=${FASTQ_DIR}/${SAMPLE}${R1_SUFFIX}
+R2_FASTQ=${FASTQ_DIR}/${SAMPLE}${R2_SUFFIX}
 export PATH=${TOOLS_DIR}/kraken2-2.0.8-beta:$PATH
 
-echo "### Trimming fastqs ### - START: $(date)"
-java -jar ${TOOLS_DIR}/Trimmomatic-0.35/trimmomatic-0.35.jar PE -phred33 -trimlog \
-    ${SAMPLE}_trimmomatic_log.txt ${FASTQ_DIR}/${SAMPLE}${R1_SUFFIX} ${FASTQ_DIR}/${SAMPLE}${R2_SUFFIX} \
-    ${SAMPLE}_trimmed${R1_SUFFIX} ${SAMPLE}_unpaired_trimmed${R1_SUFFIX} \
-    ${SAMPLE}_trimmed${R2_SUFFIX} ${SAMPLE}_unpaired_trimmed${R2_SUFFIX} \
-    ILLUMINACLIP:${TOOLS_DIR}/Trimmomatic-0.35/adapters/TruSeq3-PE-2.fa:2:30:10:2:keepBothReads \
-    LEADING:3 TRAILING:3 MINLEN:36
-echo "### Trimming fastqs ### - END: $(date)"
+if [ $SKIP_TRIMMOMATIC -eq 0 ]; then
+    UNTRIMMED_R1_FASTQ=$R1_FASTQ
+    UNTRIMMED_R2_FASTQ=$R2_FASTQ
+    R1_FASTQ=$(echo ${SAMPLE}${R1_SUFFIX} | sed "s/_R1_/_R1_trimmed_/")
+    R2_FASTQ=$(echo ${SAMPLE}${R2_SUFFIX} | sed "s/_R2_/_R2_trimmed_/")
+    UNPAIRED_R1_FASTQ=$(echo ${SAMPLE}${R1_SUFFIX} | sed "s/_R1_/_R1_trimmed_unpaired_/")
+    UNPAIRED_R2_FASTQ=$(echo ${SAMPLE}${R2_SUFFIX} | sed "s/_R2_/_R2_trimmed_unpaired_/")
+    
+    echo "### Trimming fastqs ### - START: $(date)"
+    java -jar ${TOOLS_DIR}/Trimmomatic-0.35/trimmomatic-0.35.jar PE -phred33 -trimlog \
+        ${SAMPLE}_trimmomatic_log.txt ${UNTRIMMED_R1_FASTQ} ${UNTRIMMED_R2_FASTQ} \
+        ${R1_FASTQ} ${UNPAIRED_R1_FASTQ} \
+        ${R2_FASTQ} ${UNPAIRED_R2_FASTQ} \
+        ILLUMINACLIP:${TOOLS_DIR}/Trimmomatic-0.35/adapters/TruSeq3-PE-2.fa:2:30:10:2:keepBothReads \
+        LEADING:3 TRAILING:3 MINLEN:36
+    echo "### Trimming fastqs ### - END: $(date)"
+fi
 
 echo "### Counting fastq read counts ### - START: $(date)"
-PRE_TRIM_READ_COUNT=$(echo $(zcat ${FASTQ_DIR}/${SAMPLE}${R1_SUFFIX} | wc -l ) \
-    $(zcat ${FASTQ_DIR}/${SAMPLE}${R2_SUFFIX} | wc -l) | awk '{ print ($1 + $2) / 4 }' )
-PAIR_TRIM_READ_COUNT=$(echo $(zcat ${SAMPLE}_trimmed${R1_SUFFIX} | wc -l ) \
-    $(zcat ${SAMPLE}_trimmed${R2_SUFFIX} | wc -l) | awk '{ print ($1 + $2) / 4 }' )
-UNPAIR_TRIM_READ_COUNT=$(echo $(zcat ${SAMPLE}_unpaired_trimmed${R1_SUFFIX} | wc -l ) \
-    $(zcat ${SAMPLE}_unpaired_trimmed${R2_SUFFIX} | wc -l) | awk '{ print ($1 + $2) / 4 }' )
-echo -e "sample\tread_count\ttrimmed_read_count\tunpaired_trimmed_read_count" > ${SAMPLE}.read_counts.tsv
-echo -e "$SAMPLE\t$PRE_TRIM_READ_COUNT\t$PAIR_TRIM_READ_COUNT\t$UNPAIR_TRIM_READ_COUNT" >> ${SAMPLE}.read_counts.tsv
+READ_COUNT=$(echo $(zcat $R1_FASTQ | wc -l ) \
+    $(zcat $R2_FASTQ | wc -l) | awk '{ print ($1 + $2) / 4 }' )
+echo -e "sample\tread_count" > ${SAMPLE}.read_counts.tsv
+echo -e "$SAMPLE\t$READ_COUNT" >> ${SAMPLE}.read_counts.tsv
 echo "### Counting fastq read counts ### - END: $(date)"
 
 echo "### Aligning sample to human ### - START: $(date)"
-bwa aln -t 4 $REF_FASTA ${SAMPLE}_trimmed${R1_SUFFIX} > ${SAMPLE}_R1.sai
-bwa aln -t 4 $REF_FASTA ${SAMPLE}_trimmed${R2_SUFFIX} > ${SAMPLE}_R2.sai
-bwa sampe -a 700 $REF_FASTA ${SAMPLE}_R1.sai ${SAMPLE}_R2.sai \
-    ${SAMPLE}_trimmed${R1_SUFFIX} ${SAMPLE}_trimmed${R2_SUFFIX} | \
+bwa aln -t 4 $REF_FASTA $R1_FASTQ > ${SAMPLE}_R1.sai
+bwa aln -t 4 $REF_FASTA $R2_FASTQ > ${SAMPLE}_R2.sai
+bwa sampe -a 700 $REF_FASTA ${SAMPLE}_R1.sai ${SAMPLE}_R2.sai $R1_FASTQ $R2_FASTQ | \
     samtools view -b - | samtools sort -o ${SAMPLE}_human_aligned.bam -
 samtools index ${SAMPLE}_human_aligned.bam
 echo "### Aligning sample to human ### - START: $(date)"
@@ -121,9 +128,11 @@ if [ ! -f ${SAMPLE}_no_human_vs_kraken.tsv ]; then
     echo "Final file ${SAMPLE}_no_human_vs_kraken.tsv not found. Exiting with code 1"
     exit 1
 fi
-rm ${SAMPLE}_trimmomatic_log.txt
-rm ${SAMPLE}_trimmed${R1_SUFFIX} ${SAMPLE}_trimmed${R2_SUFFIX}
-rm ${SAMPLE}_unpaired_trimmed${R1_SUFFIX} ${SAMPLE}_unpaired_trimmed${R2_SUFFIX}
+if [ $SKIP_TRIMMOMATIC -eq 0 ]; then
+    rm ${SAMPLE}_trimmomatic_log.txt
+    # rm $R1_FASTQ $R2_FASTQ
+    rm $UNPAIRED_R1_FASTQ $UNPAIRED_R2_FASTQ
+fi
 rm -r contigs_${SAMPLE} 
 rm ${SAMPLE}_contigs.fasta.amb ${SAMPLE}_contigs.fasta.ann ${SAMPLE}_contigs.fasta.bwt
 rm ${SAMPLE}_contigs.fasta.pac ${SAMPLE}_contigs.fasta.sa
